+++ src/pararnn/models/__init__.py
@@ -1,3 +1,0 @@
-from pararnn.models.xlstm_block import xLSTMBlock
-
-__all__ = ["xLSTMBlock"]